LayoutTests/fast/events/event-properties-gc-expected.txt


Testing: ErrorEvent.error
PASS event[prop] is "foo"

Testing: PopStateEvent.state
PASS event[prop] is "foo"

Testing: MessageEvent.data
PASS event[prop] is "foo"

Testing: CustomEvent.detail
PASS event[prop] is "foo"
PASS successfullyParsed is true

TEST COMPLETE
